build(karma): add browser testing capability
Note: This is not perfect unfortunately, but should allow
for the writing of tests that are executed in the browser
in an automated fashion and are working very similar to
the NodeJS based tests.

The main difference between the browser and NodeJS
tests is demonstrated by the example test case called
"get-all-method-names" in the similarly named .ts file
and is as follows:

You import the "tape" testing library, not "tap". Small
difference but actually the former is the implementation
that is able to run in the browser and in NodeJS as welll
so we have to use that and this unfortunately leads to
the negative (but temporary) side effect that the
browser tests have to be duplicated next to the NodeJS
tests. This issue will be resolved once a bigger refactor of
the build system has gone underway where we will
migrate all existing tests from "tap" to "tape" for testing
and then all tests will be browser compatible by default
as far as the test execution/assertions APIs are
concerned (of course if the component that is being
tested itself is not browser compatible then it will still
not run in browsers but that is an entirely different story).

Signed-off-by: Peter Somogyvari <peter.somogyvari@accenture.com>